third_party/blink/web_tests/fast/css/remove-class-name-expected.txt

Removing a class affected by a selector.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(document.getElementsByTagName('div')[1], null).backgroundColor is 'rgba(0, 0, 0, 0)'
PASS successfullyParsed is true

TEST COMPLETE

You should see no red.
